CrashReporter: Provide Assembly TargetFramework and Config
Having the current Runtime is not enough, if a Nuget package provides multiple binaries, it is useful to know the binary TargetFrameworkAttribute and the AssemblyConfigurationAttribute to try and identify which version is being used.

Issue: DOTNET-980

+++ CrashReporter/Crash/Dumpers/OSVersion/OSVersion.cs
@@ -162,7 +162,7 @@
                 }
             } else {
                 // WinNT
-                buildNumber = (int)((version & 0x7FFF0000) >> 16);
+                buildNumber = (version & 0x7FFF0000) >> 16;
                 PlatformId = OSPlatformId.WinNT;
             }
